WIP: Move most remaining classes to OAuth namespace
Both Google and Paratext logins still work, although the Paratext login
now does nothing with the returned access token (the handleOAuthToken
method is just a stub now). But this structure should work fine for a
Facebook OAuth login method as well.

Only class left is OAuthBase, which will be moved in the next commit.
