+++ src/Site/OAuth/GoogleOAuth.php
@@ -1,4 +1,10 @@
 <?php
+/**
+ * Created by PhpStorm.
+ * User: rmunn
+ * Date: 11/22/17
+ * Time: 1:21 PM
+ */
 
 namespace Site\OAuth;
 
@@ -6,14 +12,19 @@ use League\OAuth2\Client\Provider\AbstractProvider;
 use League\OAuth2\Client\Token\AccessToken as OAuthAccessToken;
 use Silex\Application;
 use Site\Controller\OAuthBase;
-use Symfony\Component\HttpFoundation\RedirectResponse;
+use Site\OAuth\SelectAccountOAuthProvider;
 use Symfony\Component\HttpFoundation\Request;
 
-class ParatextOAuth extends OAuthBase
+class GoogleOAuth extends OAuthBase
 {
     public function getProviderName(): string
     {
-        return "paratext";
+        return "google";
+    }
+
+    protected function handleOAuthToken(Application $app, AbstractProvider $provider, OAuthAccessToken $token)
+    {
+        return $this->loginWithOAuthToken($app, $provider, $token);
     }
 
     public function oauthCallback(Request $request, Application $app)
@@ -27,23 +38,11 @@ class ParatextOAuth extends OAuthBase
      */
     protected function getOAuthProvider($redirectUri): AbstractProvider
     {
-        $provider = new ParatextOAuthProvider([
-            'clientId' => 'DbDDp7nAdPYtuJL9L', // TODO: Move to config.php
-            'clientSecret' => '',
+        $provider = new SelectAccountOAuthProvider([
+            'clientId' => GOOGLE_CLIENT_ID,
+            'clientSecret' => GOOGLE_CLIENT_SECRET,
             'redirectUri' => $redirectUri,
         ]);
         return $provider;
-    }
-
-
-    protected function handleOAuthToken(Application $app, AbstractProvider $provider, OAuthAccessToken $token)
-    {
-        // TODO: Implement this once Paratext integration is desired
-        return new RedirectResponse("/auth/login");
-    }
-
-    public function chooseRedirectUrl(bool $tokenSuccess, Application $app) : string
-    {
-        return '/auth/show_paratext_projects';
     }
 }
